feat: full-screen add/edit note
