+++ dev-client/src/screens/AppScaffold.tsx
@@ -17,6 +17,8 @@ import {CreateProjectScreen} from 'terraso-mobile-client/screens/CreateProjectSc
 import {HomeScreen} from 'terraso-mobile-client/screens/HomeScreen';
 import {SiteTransferProjectScreen} from 'terraso-mobile-client/screens/SiteTransferProjectScreen';
 import {CreateSiteScreen} from 'terraso-mobile-client/screens/CreateSiteScreen';
+import {AddSiteNoteScreen} from 'terraso-mobile-client/components/siteNotes/AddSiteNoteScreen';
+import {EditSiteNoteScreen} from 'terraso-mobile-client/components/siteNotes/EditSiteNoteScreen';
 import {useNavigation as useNavigationNative} from '@react-navigation/native';
 import {LocationDashboardScreen} from 'terraso-mobile-client/components/sites/LocationDashboardScreen';
 import {SiteSettingsScreen} from 'terraso-mobile-client/components/sites/SiteSettingsScreen';
@@ -46,6 +48,8 @@ const screenDefinitions = {
   SITE_TEAM_SETTINGS: SiteTeamSettingsScreen,
   ADD_USER_PROJECT: AddUserToProjectScreen,
   MANAGE_TEAM_MEMBER: ManageTeamMemberScreen,
+  ADD_SITE_NOTE: AddSiteNoteScreen,
+  EDIT_SITE_NOTE: EditSiteNoteScreen,
 } satisfies ScreenDefinitions;
 
 type RootStackParamList = ParamList<typeof screenDefinitions>;
